requirements.txt
# Backend
fastapi
uvicorn[standard]
pydantic
python-multipart
pymupdf
python-dotenv

# LangChain & Embeddings
langchain
langchain-google-genai
faiss-cpu
sentence-transformers

# Google Generative AI
google-generativeai

# Frontend
streamlit
streamlit-chat
